+++ features/web3OnBoard/wallet-state/wallet-state-guards.ts
@@ -1,6 +1,7 @@
-import { isNetworkHexIdSupported, NetworkConfigHexId } from 'blockchain/networks'
+import { isNetworkHexIdSupported, NetworkConfigHexId, networkSetByHexId } from 'blockchain/networks'
 
-import { WalletManagementState } from './wallet-management-state'
+import { ensureCorrectState } from './ensure-correct-state'
+import { WalletManagementState, WalletManagementStateStatus } from './wallet-management-state'
 
 export function canTransitWithNetworkHexId(
   hexId: NetworkConfigHexId | undefined,
@@ -15,4 +16,28 @@ export function canTransitWithNetworkHexId(
   }
 
   return state.pageNetworkHexIds.includes(hexId)
+}
+
+export function shouldSendChangeNetworkOnConnected(
+  desiredNetworkHexId: NetworkConfigHexId,
+  state: WalletManagementState,
+  couldBeConnectedToTestNet: boolean,
+) {
+  ensureCorrectState(state, WalletManagementStateStatus.connected)
+  if (desiredNetworkHexId === undefined) {
+    return false
+  }
+
+  if (state.pageNetworkHexIds) {
+    return !state.pageNetworkHexIds.includes(desiredNetworkHexId)
+  }
+
+  const desiredNetworkConfig = networkSetByHexId[desiredNetworkHexId]
+
+  const possibleNetworkHexIds = [desiredNetworkConfig.mainnetHexId]
+  if (couldBeConnectedToTestNet) {
+    possibleNetworkHexIds.push(desiredNetworkConfig.testnetHexId)
+  }
+
+  return !possibleNetworkHexIds.includes(state.walletNetworkHexId)
 }
